pubhubs_hub/requirements.txt
# Taken from synapse/python_dependencies.py in the original synapse
attrs>=19.2.0,!=21.1.0
authlib>=0.14.0
Jinja2<3.1.2
typing-extensions>=3.7.4
Twisted>=18.9.0
# /use the same version as the docker image is built from
matrix-synapse==1.87
livekit-api>=0.4.1